Move location selector code into location selector module
