+++ src/components/LocationSelector.js
@@ -3,8 +3,52 @@ import FlatButton from 'material-ui/FlatButton';
 import Popover from 'material-ui/Popover';
 import Menu from 'material-ui/Menu';
 import MenuItem from 'material-ui/MenuItem';
+import { codes as isoCountryCodes } from 'iso-country-codes';
+
+const modifiedCountryNames = {
+	US: 'United States of America',
+	GB: 'United Kingdom',
+	RU: 'Russia'
+};
+
+const g20 = [
+	'AR',
+	'AU',
+	'BR',
+	'CA',
+	'CN',
+	'FR',
+	'DE',
+	'IN',
+	'ID',
+	'IT',
+	'JP',
+	'MX',
+	'RU',
+	'SA',
+	'ZA',
+	'KR',
+	'TR',
+	'GB',
+	'US'
+];
+
+const countryData = isoCountryCodes
+	.map(d => ({
+		name: modifiedCountryNames[d.alpha2] || d.name,
+		code: d.alpha2
+	}))
+	.filter(d => g20.indexOf(d.code) > -1);
 
 export default class LocationSelector extends Component {
+	countryDataByCode = code => {
+		let defaultCode = this.props.default || 'US';
+		return (
+			countryData.find(d => d.code === code) ||
+			countryData.find(d => d.code === defaultCode)
+		);
+	};
+
 	state = {
 		open: false
 	};
@@ -30,11 +74,12 @@ export default class LocationSelector extends Component {
 		});
 	};
 
-	render({ className, value, options }) {
+	render({ className, value }) {
+		let current = this.countryDataByCode(value);
 		return (
 			<div className={className}>
 				<FlatButton
-					label={value ? value.code : options[0].code}
+					label={current ? current.code : countryData[0].code}
 					onClick={this.handleClick}
 				/>
 				<Popover
@@ -48,8 +93,8 @@ export default class LocationSelector extends Component {
 						autoWidth
 						maxHeight={250}
 						onChange={this.handleSelection}
-						value={value}
-						children={options.map(option => (
+						value={current}
+						children={countryData.map(option => (
 							<MenuItem primaryText={option.name} value={option} />
 						))}
 					/>
